+++ api/src/routes/accounts.mjs
@@ -4,7 +4,7 @@ import { store } from '../lib/store.mjs'
 const Accounts = new Controller()
 
 /**
- * This method adds the authentication route to Express
+ * This method adds the accounts routes to Express
  *
  * @param {abject} app - The ExpressJS app
  */
@@ -24,30 +24,10 @@ export function routes(app) {
   /*
    * Activate account
    */
-  app.post(`${PREFIX}/activate-account`, Accounts.activateAccount)
+  app.post(`${PREFIX}/activate-account`, Accounts.activate)
 
   /*
    * Activate MFA
    */
   app.post(`${PREFIX}/activate-mfa`, Accounts.activateMfa)
-
-  /*
-   * Create an API key
-   */
-  app.post(`${PREFIX}/apikey`, Accounts.createApikey)
-
-  /*
-   * List API keys
-   */
-  app.get(`${PREFIX}/apikeys`, Accounts.listApikeys)
-
-  /*
-   * Update an API key
-   */
-  app.patch(`${PREFIX}/apikeys/:key/:action`, Accounts.updateApikey)
-
-  /*
-   * Remove an API key
-   */
-  app.delete(`${PREFIX}/apikeys/:key`, Accounts.removeApikey)
 }