+++ api/src/routes/index.mjs
@@ -3,10 +3,12 @@ import { routes as status } from '#routes/status'
 import { routes as validate } from '#routes/validate'
 import { routes as auth } from '#routes/auth'
 import { routes as accounts } from '#routes/accounts'
+import { routes as apikeys } from '#routes/apikeys'
 
 export const routes = {
   auth,
   accounts,
+  apikeys,
   core,
   status,
   validate,