+++ config/roles.mjs
@@ -1,5 +1,5 @@
 /**
- * Roles supported by this provider
+ * Roles supported by Morio
  */
 export const roles = ['user', 'manager', 'operator', 'engineer', 'root']