+++ config/services/index.mjs
@@ -4,6 +4,7 @@ import { resolveServiceConfiguration as ca } from './ca.mjs'
 import { resolveServiceConfiguration as console } from './console.mjs'
 import { resolveServiceConfiguration as connector } from './connector.mjs'
 import { resolveServiceConfiguration as core } from './core.mjs'
+import { resolveServiceConfiguration as db } from './db.mjs'
 import { resolveServiceConfiguration as dbuilder } from './dbuilder.mjs'
 import { resolveServiceConfiguration as proxy } from './proxy.mjs'
 import { resolveServiceConfiguration as ui } from './ui.mjs'
@@ -15,6 +16,7 @@ const resolvers = {
   console,
   connector,
   core,
+  db,
   dbuilder,
   proxy,
   ui,